backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
requests
pandas
numpy
yfinance
beautifulsoup4
chromadb
ollama
tqdm
pypdf
langchain
langchain-community
langchain-ollama
langchain-experimental
playwright
python-dotenv
wordcloud
